+++ lib/src/widgets/platform/mobile_osm_widget.dart
@@ -23,8 +23,8 @@ Widget getWidget({
   bool showContributorBadgeForOSM = false,
   double stepZoom = 1,
   double initZoom = 2,
-  int minZoomLevel = 2,
-  int maxZoomLevel = 18,
+  double minZoomLevel = 2,
+  double maxZoomLevel = 18,
   UserLocationMaker? userLocationMarker,
   Function(bool)? onMapIsReady,
 }) =>
